copymods: Fix typo parsing copymods= parameters in /proc/cmdline.
Normally, copymods will exit and do nothing if the target root
already contains /lib/modules/<version>.  It intended intended to allow
overriding that behavior with kernel command line argument of
string 'copymods=force'.

The parsing of that command line was simply broken.
The result was simply that it would always exit in the case that
a modules directory already existed.