overlayroot: initialize /etc/fstab if root does not have one and read-only
A root filesystem may not have an /etc/fstab.  That could be for either:
a.) not needed - the kernel command line has root mount options and no
  other filesystems are mounted.
b.) system has systemd.mounts - this is currently not supported by
overlayroot.

If the kernel command line contains 'ro', then overlayroot will
correctly leave root mounted 'ro'.  The problem was that without an
/etc/fstab, systemd would not remount the root filesystem rw, and
thus failure would ensue.

The change here is to initialize an fstab in the case that the root
filesystem did not contain one, *and* 'ro' is present on the cmdline.
In cases where 'ro' is not present, then overlayroot will leave /
mounted as rw, so systemd does not need to remount rw.

LP: #1890803