overlayroot: overwrite existing fstab when overlayrootifying fstab
The previous fix for handling missing fstab refactored out the initial
write to ${ROOTMNT}/etc/fstab which emitted the overlayroot fstab
header (clobbering any existing fstab that was in the rootfs).  Now
the header is generated in overlayrootify_fstab; and the output was
*appended* which worked when the write was in two stages (header +
contents);  Now that the header is generated with the content the
output must clobber the target file.